# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "noplagiat-analytics API",
    "description": "Antiplagiarism analytics dashboard - Toraighyrov University",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Sign in with a local account.",
        "description": "Every failure - unknown name, deactivated account, no password set, wrong\npassword - is one `401` with one message. Distinguishing them would turn the\nendpoint into an oracle for which login names exist, and the operator who\ncan create accounts is not the person doing the guessing.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "session established",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "the credentials were not accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "a field was empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "too many sign-in attempts from this address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "End the current session and clear the cookie.",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "session ended",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "401": {
            "description": "no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "missing or mismatched CSRF token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The current session's identity and authorization.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "description": "One rejected query parameter or body field.",
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string",
            "description": "Parameter name, or a `garde` path such as `to` or `items[0].code`."
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "description": "Sign-in request.\n\nNo name and no e-mail: those are account fields an operator sets with the\nCLI, so this endpoint cannot be used to put a person's name into the\ndatabase (AGENTS.md invariant #1).",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "The account's password. Never logged, never echoed."
          },
          "username": {
            "type": "string",
            "description": "Login name, matched case-insensitively."
          }
        },
        "additionalProperties": false
      },
      "LogoutResponse": {
        "type": "object",
        "required": [
          "next_path"
        ],
        "properties": {
          "next_path": {
            "type": "string",
            "description": "Where the browser should go once the session is gone. Always the sign-in\npage: with no identity provider there is no second session to end."
          }
        }
      },
...
      "MeResponse": {
        "type": "object",
        "required": [
          "username",
          "roles",
          "request_access_path",
          "csrf_token"
        ],
        "properties": {
          "csrf_token": {
            "type": "string",
            "description": "Hex CSRF token for this session - send it as `x-csrf-token` on every\nmutating request."
          },
          "request_access_path": {
            "type": "string",
            "description": "Where to send an account that holds no grants (TZ §5)."
          },
          "role": {
            "type": [
              "string",
              "null"
            ],
            "description": "Widest role held, or `null` for an account awaiting access."
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleGrantDto"
            }
          },
          "scope": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ScopeDto",
                "description": "Visibility the roles grant, or `null` when none of them reaches the\ninternal contour."
              }
            ]
          },
          "username": {
            "type": "string",
            "description": "Login name of the signed-in account."
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "One RFC 7807 problem document.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "description": "Explanation specific to this occurrence."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "description": "Per-field detail for 422 responses."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Request path this problem occurred on."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary - the same for every occurrence of a type."
          },
          "type": {
            "type": "string",
            "description": "Stable problem-type identifier, e.g. `/problems/validation-failed`."
          }
        }
      },
...
      "RoleGrantDto": {
        "type": "object",
        "description": "One `(role, scope)` grant, as `/api/auth/me` reports it.",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "type": "string",
            "description": "`staff`, `dept_head`, `dean`, `ethics`, `compliance` or `admin`."
          },
          "scope_department_code": {
            "type": [
              "string",
              "null"
            ],
            "description": "Dictionary code of the scoped department; see `scope_faculty_code`."
          },
          "scope_department_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "scope_faculty_code": {
            "type": [
              "string",
              "null"
            ],
            "description": "Dictionary code of the scoped faculty, so a client can name the unit a\ndean governs without resolving an opaque id. `null` for an unscoped\ngrant, and for an id whose dictionary row has been deleted."
          },
          "scope_faculty_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
...
      "ScopeDto": {
        "type": "object",
        "description": "The caller's RBAC visibility, as `compliance::Scope` renders on the wire.\n\nCarries ids rather than codes because that is what the scope *is*; the UI\nresolves them through the dictionary endpoints.",
        "required": [
          "kind"
        ],
        "properties": {
          "department_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "faculty_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "kind": {
            "type": "string",
            "description": "`all`, `faculty` or `department`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "np_session",
        "description": "Opaque session id, HttpOnly + Secure + SameSite=Lax"
      }
    }
  },
...
}